removing bitcoinjprovider package and moving classes up level up in the hiearchy.
